+++ 000_Randomness/NormalDistributionScene.h
@@ -11,7 +11,7 @@ public:
   NormalDistributionScene(SceneStack& sceneStack,int winWidth, int winHeight, float bottomPadding = 10.0f, float xPadding = 10.f, float mean = 5.0, float stddev = 2.0, int samples = 10000);
   
   virtual void Init() override;
-  virtual void Update(uint32_t dt) override;
+  virtual void Update(double dt) override;
   virtual void Draw() override;
   virtual const std::string& GetSceneName() const override;
   //virtual const void HandleInput() override;
