+++ 000_Randomness/PerlinNoiseScene.h
@@ -11,7 +11,7 @@ public:
   PerlinNoiseScene(SceneStack& sceneStack, int winWidth, int winHeight, float bottomPadding = 10.0f, float xPadding = 10.f, float factor = .5);
   
   virtual void Init() override;
-  virtual void Update(uint32_t dt) override;
+  virtual void Update(double dt) override;
   virtual void Draw() override;
   virtual const std::string& GetSceneName() const override;
   //virtual const void HandleInput() override;
