+++ 001_Forces/ForcesScene.h
@@ -13,7 +13,7 @@ public:
   ForcesScene(SceneStack& sceneStack, int winWidth, int winHeight);
 
   virtual void Init() override;
-  virtual void Update(uint32_t dt) override;
+  virtual void Update(double dt) override;
   virtual void Draw() override;
   virtual const std::string& GetSceneName() const override;
   virtual const void HandleInput() override;
