+++ 001_Forces/VelocityScene.h
@@ -11,7 +11,7 @@ public:
   VelocityScene(SceneStack& sceneStack, int winWidth, int winHeight, float bottomPadding = 10.0f, float xPadding = 10.f, float maxVelocity = 10.0f, Vector2 position = Vector2{0,0}, Vector2 velocity = Vector2{0,0}, Vector2 acceleration = Vector2{-0.01, 0.01});
   
   virtual void Init() override;
-  virtual void Update(uint32_t dt) override;
+  virtual void Update(double dt) override;
   virtual void Draw() override;
   virtual const std::string& GetSceneName() const override;
   virtual const void HandleInput() override;
